feat(ps): add option to backup and simulate the process for git deployment

+++ powershell/Deploy-GitCommit.ps1
@@ -7,11 +7,18 @@ function Deploy-GitCommit {
 
 		[string]
 		[Parameter(Mandatory=$true)]
-		$SourceFolder,
+		$DestinationFolder,
 
 		[string]
-		[Parameter(Mandatory=$true)]
-		$DestinationFolder
+		[Parameter(Mandatory=$false)]
+		$SourceFolder = (Get-Location).Path,
+
+		[string]
+		[Parameter(Mandatory=$false)]
+		$BackupFolder,
+
+		[switch]
+		$Simulate
 	)
 
 	Begin {
@@ -20,12 +27,20 @@ function Deploy-GitCommit {
 			Write-Error "Git command not found"
 		}
 
+		$OldBaseDirectory = (Get-Location).Path
+		$OldGitBranch = Get-GitCurrentBranch
 		$SourceFolder = (Get-Item $SourceFolder).FullName.Trim('\')
+		Set-Location $SourceFolder
 		$DestinationFolder = (Get-Item $DestinationFolder).FullName.Trim('\')
-		$BaseDirectory = (Get-Location).Path.Trim('\')
 		Write-Verbose "Deploying from '$SourceFolder' to '$DestinationFolder'"
 		if ([string]::IsNullOrWhiteSpace($SourceFolder) -or [string]::IsNullOrWhiteSpace($DestinationFolder)) {
 			Write-Error "Source/destination folder can't be empty'"
+		}
+
+		Write-Verbose "Switching to git commit $CommitId from branch $OldGitBranch"
+		if (-not $Simulate) {
+			git stash | Out-Null
+			git checkout $CommitId | Out-Null
 		}
 	}
 
@@ -33,20 +48,83 @@ function Deploy-GitCommit {
 		$Files = &$GitCommand show --name-only --oneline --pretty="format:" $CommitId
 		if ($null -eq $Files -or $Files.Length -eq 0) {
 			Write-Warning "No files found in git commit"
+		} else {
+			$InlineFiles = Join-String -InputObject $Files -Separator "`n- "
+			Write-Verbose "For the following files:`n- $InlineFiles"
 		}
+
+		if (-not([string]::IsNullOrWhiteSpace($BackupFolder))) {
+			if (-not(Test-Path -Path $BackupFolder -PathType Container) -and -not $Simulate) {
+				New-Item -Path $BackupFolder -ItemType Directory -Force | Out-Null
+			}
+			if (Test-Path -Path $BackupFolder -PathType Container) {
+				$BackupFolder = (Get-Item $BackupFolder).FullName.Trim('\')
+			}
+			$PreFolder = Join-Path $BackupFolder -ChildPath 'pre'
+			Write-Verbose "Creating backup pre folder: $PreFolder"
+			if (-not $Simulate) {
+				New-Item -ItemType Directory -Path $PreFolder -Force | Out-Null
+			}
+			Copy-FilesInner -Files $Files -BaseDirectory $SourceFolder -Simulate:$Simulate `
+				-DestinationFolder $PreFolder -SourceFolder $DestinationFolder
+			$PostFolder = Join-Path $BackupFolder -ChildPath 'update'
+			Write-Verbose "Creating backup post folder: $PostFolder"
+			if (-not $Simulate) {
+				New-Item -ItemType Directory -Path $PostFolder -Force | Out-Null
+			}
+			Copy-FilesInner -Files $Files -BaseDirectory $SourceFolder -Simulate:$Simulate `
+				-DestinationFolder $PostFolder -SourceFolder $SourceFolder
+		}
+		Copy-FilesInner -Files $Files -BaseDirectory $SourceFolder -Simulate:$Simulate `
+			-DestinationFolder $DestinationFolder -SourceFolder $SourceFolder
+	}
+
+	End {
+		if (-not $Simulate) {
+			git checkout $OldGitBranch | Out-Null
+			git stash pop | Out-Null
+		}
+		Set-Location $OldBaseDirectory
+	}
+}
+
+function Get-GitCurrentBranch {
+	git branch | ?{ $_ -match '^\*' } | %{ $_.Trim('*').Trim() }
+}
+
+function Copy-FilesInner {
+	[CmdletBinding()]
+	param(
+		[string[]]$Files,
+		[string]$BaseDirectory,
+		[string]$DestinationFolder,
+		[string]$SourceFolder,
+		[switch]$Simulate
+	)
+
+	Process {
 		foreach ($File in $Files) {
-			$RelativeDirectory = (Get-Item -Path $File).Directory.FullName.Replace($BaseDirectory, [string]::Empty).TrimStart('\')
+			$Item = Get-Item -Path $File -ErrorAction SilentlyContinue
+			$RelativeDirectory = $null -ne $Item ? $Item.Directory.FullName.Replace($BaseDirectory, [string]::Empty).TrimStart('\') : $File.Replace('/', '\')
 			$TargetFolder = $DestinationFolder
 			$SourcePath = Join-Path -Path $SourceFolder -ChildPath $File
 			if (-not([string]::IsNullOrWhiteSpace($RelativeDirectory))) {
 				$TargetFolder = Join-Path $DestinationFolder -ChildPath $RelativeDirectory
 				if (-not(Test-Path -PathType Container -Path $TargetFolder)) {
-					Write-Verbose "Creating $TargetFolder"
-					New-Item -ItemType Directory -Path $TargetFolder | Out-Null
+					Write-Verbose "Creating folder: $TargetFolder"
+					if (-not $Simulate) { 
+						New-Item -ItemType Directory -Path $TargetFolder | Out-Null
+					}
 				}
 			} 
-			Write-Verbose "Copying '$SourcePath' to '$TargetFolder'"
-			Copy-Item -Path $SourcePath -Destination $TargetFolder | Out-Null
+			if (Test-Path -Path $SourcePath -PathType Leaf) {
+				Write-Verbose "Copying '$SourcePath' to '$TargetFolder'"
+				if (-not $Simulate) {
+					Copy-Item -Path $SourcePath -Destination $TargetFolder | Out-Null
+				}
+			} else {
+				Write-Warning "The file '$SourcePath' doesn't exist, skipping"
+			}
 		}
 	}
 }
